split python code into modules
